fix: 修复 CoC DM fallback 骰面文案
